refactor: move RootStore out of index
