+++ src/stores/Howto/howto.store.test.ts
@@ -11,13 +11,13 @@ import { FactoryUser } from 'src/test/factories/User'
 import { HowtoStore } from './howto.store'
 
 import type { IHowtoDB, IUser } from 'src/models'
-import type { RootStore } from '..'
+import type { IRootStore } from '../RootStore'
 
 const factory = async (
   howTos: IHowtoDB[] = [FactoryHowto({})],
   userOverloads?: Partial<IUser>,
 ) => {
-  const store = new HowtoStore({} as RootStore)
+  const store = new HowtoStore({} as IRootStore)
 
   const howToItem = howTos[0]
