+++ src/stores/RootStore.tsx
@@ -13,35 +13,9 @@ import { UserNotificationsStore } from './User/notifications.store'
 import { UserStore } from './User/user.store'
 import { DatabaseV2 } from './databaseV2'
 
-export class RootStore {
-  dbV2 = new DatabaseV2()
-  stores = stores(this)
-}
-
-// the following stores are passed into a top level app provider and can be accessed through @inject
-// all stores are also shared a top-level root store, which provides access to the main database and
-// all other stores if required. More info on this pattern can be found at: https://mobx.js.org/best/store.html
-
-// NOTE - As all stores are injected at the same time it is best to avoid using many constructor methods
-// as these will be called immediately, and instead use init() or similar methods that can be called
-// from a page (see common/module store for example)
-const stores = (rootStore: RootStore) => {
-  const stores: IStores = {
-    aggregationsStore: new AggregationsStore(rootStore),
-    howtoStore: new HowtoStore(rootStore),
-    userStore: new UserStore(rootStore),
-    tagsStore: new TagsStore(rootStore),
-    categoriesStore: new CategoriesStore(rootStore),
-    researchCategoriesStore: new ResearchCategoriesStore(rootStore),
-    mapsStore: new MapsStore(rootStore),
-    messageStore: new MessageStore(rootStore),
-    themeStore: new ThemeStore(),
-    userNotificationsStore: new UserNotificationsStore(rootStore),
-    questionStore: new QuestionStore(rootStore),
-    questionCategoriesStore: new QuestionCategoriesStore(rootStore),
-    discussionStore: new DiscussionStore(rootStore),
-  }
-  return stores
+export interface IRootStore {
+  dbV2: DatabaseV2
+  stores: IStores
 }
 
 export interface IStores {
@@ -58,4 +32,37 @@ export interface IStores {
   questionStore: QuestionStore
   questionCategoriesStore: QuestionCategoriesStore
   discussionStore: DiscussionStore
+}
+
+export class RootStore implements IRootStore {
+  dbV2 = new DatabaseV2()
+  stores = stores(this)
+}
+
+// The following stores are passed into a top level app provider and can be accessed through @inject
+// all stores are also shared a top-level root store, which provides access to the main database and
+// all other stores if required.
+// More info on this pattern can be found at:
+// https://mobx.js.org/defining-data-stores.html#combining-multiple-stores
+
+// NOTE - As all stores are injected at the same time it is best to avoid using many constructor methods
+// as these will be called immediately, and instead use init() or similar methods that can be called
+// from a page (see common/module store for example)
+const stores = (rootStore: IRootStore) => {
+  const stores: IStores = {
+    aggregationsStore: new AggregationsStore(rootStore),
+    howtoStore: new HowtoStore(rootStore),
+    userStore: new UserStore(rootStore),
+    tagsStore: new TagsStore(rootStore),
+    categoriesStore: new CategoriesStore(rootStore),
+    researchCategoriesStore: new ResearchCategoriesStore(rootStore),
+    mapsStore: new MapsStore(rootStore),
+    messageStore: new MessageStore(rootStore),
+    themeStore: new ThemeStore(),
+    userNotificationsStore: new UserNotificationsStore(rootStore),
+    questionStore: new QuestionStore(rootStore),
+    questionCategoriesStore: new QuestionCategoriesStore(rootStore),
+    discussionStore: new DiscussionStore(rootStore),
+  }
+  return stores
 }
